+++ StenGine/src/Graphics/OpenGL/GLTexture.cpp
@@ -8,7 +8,7 @@ namespace StenGine
 GLTexture::GLTexture(uint32_t width, uint32_t height, void* tex)
 	: m_width(width)
 	, m_height(height)
-	, m_texture(reinterpret_cast<GLuint>(tex))
+	, m_texture(static_cast<GLuint>(reinterpret_cast<intptr_t>(tex)))
 {
 	m_textureHandle = glGetTextureHandleARB(m_texture);
 	glMakeTextureHandleResidentNV(m_textureHandle);
